requirements.txt
flask
flask-pydantic
sqlalchemy
gunicorn
psycopg2-binary
loguru
